gpu: nvgpu: fix error condition in nvgpu_safe_mult_u64
This patch fixes error condition in nvgpu_safe_mult_u64. As this
function is operating on u64, ULONG_MAX should be used to check
error condition.

NVGPU-3432

Reviewed-on: https://git-master.nvidia.com/r/2128612
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/safe_ops.h
@@ -97,7 +97,7 @@ static inline u64 nvgpu_safe_mult_u64(u64 ul_a, u64 ul_b)
 {
 	if (ul_a == 0 || ul_b == 0) {
 		return 0UL;
-	} else if (ul_a > UINT_MAX / ul_b) {
+	} else if (ul_a > ULONG_MAX / ul_b) {
 		BUG();
 	} else {
 		return ul_a * ul_b;
